chore: factor-out to abstract
